backend/requirements.txt
# coin-tracker/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
pydantic[email]
httpx